Backend/requirements.txt
flask==2.2.3
flask-socketio==5.3.2
flask-cors==3.0.10
gunicorn==20.1.0
eventlet==0.33.3
psycopg2-binary==2.9.5
pandas==1.5.3
bcrypt==4.0.1
python-dotenv==1.0.0
yagmail==0.15.293
requests==2.28.2
google-auth==2.16.2
numpy==1.24.2
